// src/plugins/ws/protobuf/ts/proto/OptionOrder.ts
export interface OptionOrder {
    /**
     * @generated from protobuf field: repeated OptionOrder.Body optionOrder = 1;
     */
    optionOrder: OptionOrder_Body[];
}
/**
 * @generated from protobuf message OptionOrder.Body
 */
export interface OptionOrder_Body {
    /**
     * @generated from protobuf field: sint64 uid = 1;
     */
    uid: string;
    /**
     * @generated from protobuf field: sint64 orderId = 2;
     */
    orderId: string;
    /**
     * @generated from protobuf field: sint64 businessId = 3;
     */
    businessId: string;
    /**
     * @generated from protobuf field: string orderState = 4;
     */
    orderState: string;
    /**
     * @generated from protobuf field: sint64 optionId = 5;
     */
    optionId: string;
    /**
     * @generated from protobuf field: string symbol = 6;
     */
    symbol: string;
    /**
     * @generated from protobuf field: int32 profitable = 7;
     */
    profitable: number;
    /**
     * @generated from protobuf field: string profitValue = 8;
     */
    profitValue: string;
    /**
     * @generated from protobuf field: string coinSymbol = 9;
     */
    coinSymbol: string;
}
// @generated message type with reflection information, may provide speed optimized methods
class OptionOrder$Type extends MessageType<OptionOrder> {
    constructor() {
        super("OptionOrder", [
            { no: 1, name: "optionOrder", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => OptionOrder_Body }
        ]);
    }
    create(value?: PartialMessage<OptionOrder>): OptionOrder {
        const message = { optionOrder: [] };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<OptionOrder>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: OptionOrder): OptionOrder {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* repeated OptionOrder.Body optionOrder */ 1:
                    message.optionOrder.push(OptionOrder_Body.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: OptionOrder, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* repeated OptionOrder.Body optionOrder = 1; */
        for (let i = 0; i < message.optionOrder.length; i++)
            OptionOrder_Body.internalBinaryWrite(message.optionOrder[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message OptionOrder
 */
export const OptionOrder = new OptionOrder$Type();
// ...
export const OptionOrder_Body = new OptionOrder_Body$Type();